Add 4 new functions to work with steering data
  H5FDdsm_EXPORT herr_t H5FD_dsm_steering_begin_query();
  H5FDdsm_EXPORT herr_t H5FD_dsm_steering_end_query();
  H5FDdsm_EXPORT herr_t H5FD_dsm_steering_gethandle(const char *name, hid_t *handle);
  H5FDdsm_EXPORT herr_t H5FD_dsm_steering_freehandle(hid_t handle);

begin and end manage a cache of handles used by the steering queries to reduce
the need to open cloase the file and groups.

gethandle and freehandle are provided to access a large dataset directly.
Once retrieved, the user may access the dataset using hyperslabs like
any conventional HDF5 dataset.