Add H5FD_dsm_steering_vector_set to allow Interactions to be written to by the simulation
Simulation can write variables equivalently to the steering end so that simpler
flags and parameters may be exchanged prior to any main data exchange.

Clean up the cached use of hid_t file/fapl/group ids for the interactions
so that the same code is used for read/writing by steerer, or read/write by simulation.