Add GetSteeringValues which can be set/Get from either side.
This addition (+Fortran wrappers) is provided to allow the simulation to export simple parameters
which can be read back by the GUI (GetSteeringValues). The params are generated
using the proxy helper. Currently we are using this to export TimeeRange from SPH-flow
so that ParaView can setup animation type properties on objects or parameters.