Fixes for static communicator usage with ParaView
Do write config when using static communicator
Make sure host name is initialized to empty string to avoind garbage in config

When publish is called using static communicator, we call MPI_Barrier
using MPI_COMM_WORLD to allow the simulation to resume. This needs to be
cleaned up/standardized within the steering/other init so that it is handled
automatically rather than by the user manually adding a Barrier.