Move some cache logic out of Steerer and into DSMManager
Add a collective Open/Close command so that Apps like Paraview can
Force a parallel open before some method tries to open on a single node.

When opening in ReadOnly mode from the steerer, we open in serial mode
to help with the problem of ParaView only calling the function on a single
node and not on all.

Needs to be used with fixes from Icarus and Xdmf.