Clean up Open/Close DSM and use a stack to track open close pairs
Each time the DSM is opened, we push the mode onto the stack,
on the first open, the file is opened.
Each time close is called, pop one off the stack and when the
stack is empty we close the file.

An open in read only mode is allowed if the file is already open in read/write mode,
but not vice versa.

We handle the interaction cache in the same way, all open/close calls go via the
DSMmanager and leave the management of file handles to the manager, just handle the
interaction group.

Fixes to the Cache handling for serial/parallel mode

If not specified and the dsm is already open, then the BeginInteraction
opens the dsm in the same mode that the dsm is already open.

Remove Exception and replace with error message